add users admin page

+++ app/Http/Controllers/Admin/UserController.php
@@ -2,7 +2,9 @@
 
 namespace Convoy\Http\Controllers\Admin;
 
+use Convoy\Http\Controllers\ApplicationApiController;
 use Convoy\Http\Controllers\Controller;
+use Convoy\Http\Requests\Admin\Users\UpdateUserRequest;
 use Convoy\Http\Requests\Admin\Users\StoreUserRequest;
 use Convoy\Models\Filters\FiltersUser;
 use Convoy\Models\User;
@@ -11,8 +13,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Spatie\QueryBuilder\AllowedFilter;
 use Spatie\QueryBuilder\QueryBuilder;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
-class UserController extends Controller
+class UserController extends ApplicationApiController
 {
     public function index(Request $request)
     {
@@ -24,6 +27,13 @@ class UserController extends Controller
         return fractal($users, new UserTransformer())->respond();
     }
 
+    public function show(User $user)
+    {
+        $user->loadCount(['servers']);
+
+        return fractal($user, new UserTransformer())->respond();
+    }
+
     public function store(StoreUserRequest $request)
     {
         $user = User::create([
@@ -31,15 +41,35 @@ class UserController extends Controller
             'email' => $request->email,
             'password' => Hash::make($request->password),
             'root_admin' => $request->root_admin,
+        ])->loadCount(['servers']);
+
+        return fractal($user, new UserTransformer())->respond();
+    }
+
+    public function update(UpdateUserRequest $request, User $user)
+    {
+        $user->update([
+            'name' => $request->name,
+            'email' => $request->email,
+            'root_admin' => $request->root_admin,
+            ...(is_null($request->password) ? [] : ['password' => Hash::make($request->password)])
         ]);
 
-        return redirect()->route('admin.users.show', [$user->id]);
+        $user->loadCount(['servers']);
+
+        return fractal($user, new UserTransformer())->respond();
     }
 
     public function destroy(User $user)
     {
+        $user->loadCount('servers');
+
+        if ($user->servers_count > 0) {
+            throw new BadRequestHttpException('The user cannot be deleted with servers still associated.');
+        }
+
         $user->delete();
 
-        return redirect()->route('admin.users');
+        return $this->returnNoContent();
     }
 }
